Sort tags alphabetically so rendered docs order is consistent
The rendered documentation orders sections by the top-level `tags:` array,
not by path order, so newly added tags appeared at the end. Generalize the
sort helper (now utils/sortSpec.js) to also sort the tags array by name, and
apply it. Update package.json (`yarn sort`, validate --check) accordingly.
